requirements.txt
python-dotenv
kaggle